+++ src/components/select/select-multiple/select-multiple.vue
@@ -119,7 +119,8 @@
                   ]"
                 >
                   <Icon v-if="props.clearable && inputText" icon="ph:x" @click.stop="handleClear" />
-                  <Icon icon="ph:caret-down" />
+                  <Icon v-if="props.inputLoader" icon="svg-spinners:270-ring" />
+                  <Icon v-else icon="ph:caret-down" />
                 </div>
               </template>
 
@@ -159,7 +160,7 @@
             :menu-list="multiSelectOptions"
             :group-items-by="props.groupItemsBy"
             :pre-selected-items="Array.isArray(multiSelectModel) ? multiSelectModel.flat() : [multiSelectModel]"
-            :loading="props.loading"
+            :loading="props.optionsLoader || props.loading"
             :item-icon="props.itemIcon"
             :lozenge="props.lozenge"
             :supporting-display-text="props.supportingDisplayText"